Move Scene.interact() objects from data_structures.py into scene.py
