+++ manim/data_structures.py
@@ -29,45 +29,3 @@ class MethodWithArgs:
     method: MethodType
     args: Iterable[Any]
     kwargs: dict[str, Any]
-
-
-@dataclass
-class SceneInteractContinue:
-    """Object which, when encountered in :meth:`.Scene.interact`, triggers
-    the end of the scene interaction, continuing with the rest of the
-    animations, if any. This object can be queued in :attr:`.Scene.queue`
-    for later use in :meth:`.Scene.interact`.
-
-    Attributes
-    ----------
-    sender : str
-        The name of the entity which issued the end of the scene interaction,
-        such as ``"gui"`` or ``"keyboard"``.
-    """
-
-    __slots__ = ["sender"]
-
-    sender: str
-
-
-class SceneInteractRerun:
-    """Object which, when encountered in :meth:`.Scene.interact`, triggers
-    the rerun of the scene. This object can be queued in :attr:`.Scene.queue`
-    for later use in :meth:`.Scene.interact`.
-
-    Attributes
-    ----------
-    sender : str
-        The name of the entity which issued the rerun of the scene, such as
-        ``"gui"``, ``"keyboard"``, ``"play"`` or ``"file"``.
-    kwargs : dict[str, Any]
-        Additional keyword arguments when rerunning the scene. Currently,
-        only ``"from_animation_number"`` is being used, which determines the
-        animation from which to start rerunning the scene.
-    """
-
-    __slots__ = ["sender", "kwargs"]
-
-    def __init__(self, sender: str, **kwargs: Any) -> None:
-        self.sender = sender
-        self.kwargs = kwargs
